[DEL] renamed Sources.cs in User.cs (containing all the API functions under the User section) for being consistent with the API description. Deleted Sources.cs [ADD] added User.cs and missing API functions implementation: GetUser, UpdateUser, GetInvoices, GetMessagingStats, GetSpendingStats, AuthenticateUser, RefreshToken, GetSingleSubaccount, CloseSubaccount, GetSubAccounts, InviteSubAccount,
